Correct ccall signatures and drop the dead info handler
The four readstat_set_*_handler calls declared a return type of Int,
which is 64 bits, where C returns readstat_error_t, a 4-byte enum. The
values are discarded, so this was harmless in practice, but the
declaration was wrong.

readstat_get_modified_time returns time_t, which is signed. Declaring it
UInt turned any pre-1970 timestamp into a huge positive number that
unix2datetime then choked on.

handle_info! was compiled into a cfunction on every parse and then never
registered: readstat dropped readstat_set_info_handler long ago, and the
metadata handler already sets rows and columns.

+++ src/C_interface.jl
@@ -4,7 +4,7 @@ function readstat_get_file_label(metadata::Ptr{Nothing})
 end
 
 function readstat_get_modified_time(metadata::Ptr{Nothing})
-    return ccall((:readstat_get_modified_time, libreadstat), UInt, (Ptr{Nothing},), metadata)
+    return ccall((:readstat_get_modified_time, libreadstat), Int64, (Ptr{Nothing},), metadata)
 end
 
 function readstat_get_file_format_version(metadata::Ptr{Nothing})
